Move ListenerRegistrationImpl to core
